Refactoring: Separate module codes into `module` directory
IoT.js-DCO-1.0-Signed-off-by: Youngil Choi duddlf.choi@samsung.com
