+++ src/module/iotjs_module_buffer.h
